Copied Current DCAL Repo Over
Considering whether I move step 5 forward, and transfer this repo to opendatacube now, and then re-vector the web links to DCAL to that repo. Or hold off on that for now. Issue is these current notebooks aren't necessarily validated, and so may not 'work' with current ODC and so we don't want to put them in the official repo. TBD.